refactor(surface): move the projection vocabulary down to @kolu/surface/verbs
`BespokeTool`, `toolName` and the Effect Schema → JSON-Schema bridge lived in
`@kolu/surface-mcp`, and the name was the only MCP-specific thing about any of
them: a verb record is what an app hands to EVERY projecting face, a flat name
is what a face with no dots to spend calls a procedure, and the bridge is how a
schema-less caller learns what an input looks like. A second face — the argv
projection landing next — would otherwise import the MCP SDK to obtain a type
and compute names by its own grammar, which is exactly the drift
`@kolu/surface/expose` was written to end for the allowlist.

So they move to `packages/surface/src/verbs.ts` (`@kolu/surface/verbs`):
`SurfaceVerb` (was `BespokeTool`), `SurfaceVerbInputSchema` (was
`ToolInputSchema`), `toolName`, `toInputSchema`/`inputSchema`, and the wrapping
rule whose private `value` key all three of its moves must agree on.
`@kolu/surface-mcp` re-exports the old names from its root, so `kolu-mcp` and
every other consumer keeps compiling unchanged.

No behaviour change. `surface-mcp`'s four remaining test files are untouched and
green (78 tests), which is the proof; `jsonschema.test.ts` moved to
`packages/surface/src/verbs.test.ts` with its assertions verbatim and only its
import repointed, plus new pins for `toolName` and the wrapping rule now that
both are public framework API.

Reference docs move with the API: `ref-surface.mdx` gains the projection-
vocabulary section, `ref-surface-mcp.mdx` says which of its exports are now
re-exports.
