+++ packages/surface-mcp/src/server.ts
@@ -51,18 +51,9 @@ import {
 import { Effect, Option, Schema, Stream } from "effect";
 import { match } from "ts-pattern";
 import { COLLECTION_PREFIX, type ResourceEntry, resolveExpose } from "./expose";
-import { inputSchema } from "./jsonschema";
 import { type PusherConnection, ResourcePusher } from "./pusher";
-import {
-  type BespokeTool,
-  brand,
-  fail,
-  failFrom,
-  messageOf,
-  ok,
-  type ToolResult,
-} from "./tools";
-import { unwrapArgs } from "./wrapping";
+import { brand, fail, failFrom, messageOf, ok, type ToolResult } from "./tools";
+import { inputSchema, type SurfaceVerb, unwrapArgs } from "@kolu/surface/verbs";
 
 /** The structural shape of a served-surface client the adapter needs. The
  *  concrete client is what `buildSurfaceFace` mints (`surfaceClientRef`, the
@@ -111,8 +102,10 @@ export interface ServeSurfaceAsMcpOptions<S extends SurfaceSpec> {
   client: () => ClientOrConnection | Promise<ClientOrConnection>;
   /** Default-deny allowlist — what an agent may touch. */
   expose: ExposeMap<S>;
-  /** Hand-authored, call-shaped MCP tools composing over the live client. */
-  tools?: Record<string, BespokeTool>;
+  /** Hand-authored, call-shaped verbs composing over the live client — the
+   *  framework record (`SurfaceVerb`), so the same table also projects as argv
+   *  through `@kolu/surface-cli`. */
+  tools?: Record<string, SurfaceVerb>;
   serverInfo?: { name: string; version: string };
   /** The server's own `instructions`, answered to a host at `initialize` — where
    *  an embedding app teaches an agent the domain the surface is about ("a node
@@ -148,7 +141,7 @@ export async function serveSurfaceAsMcp<S extends SurfaceSpec>(
   // the full pass each time.
   const bespokeTools = new Map<
     string,
-    { tool: BespokeTool; schema: Record<string, unknown>; wrapped: boolean }
+    { tool: SurfaceVerb; schema: Record<string, unknown>; wrapped: boolean }
   >(
     Object.entries(bespoke).map(([name, t]) => [
       name,
@@ -518,7 +511,7 @@ export async function serveSurfaceAsMcp<S extends SurfaceSpec>(
       title: tool.title,
       description: tool.description,
       inputSchema: schema,
-      // Conservative default (see `BespokeTool.mutates`): an absent `mutates`
+      // Conservative default (see `SurfaceVerb.mutates`): an absent `mutates`
       // is treated as MUTATING, so an unannotated tool is never advertised as
       // auto-approvable read-only. A genuinely read-only tool opts in with an
       // explicit `mutates: false`.
