+++ packages/surface-mcp/src/tools.ts
@@ -4,10 +4,13 @@
  *   - An **exposed procedure** (`expose: { "node.rerun": { tool: ... } }`):
  *     dispatch calls `client.surface[ns][verb](args)` and wraps the result.
  *   - A **bespoke tool** (`tools: { run: { input, handler } }`): a
- *     hand-authored MCP tool whose handler composes over the live client.
- *     Genuinely call-shaped capabilities (spawn-and-await, path-guarded
- *     reads) ride here, sharing the package's Schema→JSON-Schema + result
- *     framing spine — they just supply an Effect Schema input and a function.
+ *     hand-authored, call-shaped verb whose handler composes over the live
+ *     client. Genuinely call-shaped capabilities (spawn-and-await, path-guarded
+ *     reads) ride here, sharing the framework's Schema→JSON-Schema bridge and
+ *     this package's result framing — they just supply an Effect Schema input
+ *     and a function. The record itself is the framework's `SurfaceVerb`
+ *     (`@kolu/surface/verbs`), re-exported from this package's root as
+ *     `BespokeTool`: an app hands the SAME record to the CLI face.
  *
  * Both wrap their return as a `ToolResult` — the model's prose in `content`,
  * the caller's data in `structuredContent` — and surface a thrown error as
@@ -37,47 +40,7 @@
  * parameter and let interruption do the work.
  */
 
-import type { Effect, Schema } from "effect";
-import { wrapValue } from "./wrapping";
-
-/** A bespoke tool's input schema: any context-free Effect Schema whose DECODED
- *  type is the handler's `args`. The same bound `@kolu/surface`'s `WireSchema<T>`
- *  puts on every spec schema — `RD`/`RE` pinned to `never`, because decoding an
- *  MCP tool argument has no Effect environment to draw services from. */
-export type ToolInputSchema<I> = Schema.Codec<I, unknown, never, never>;
-
-/** A hand-authored MCP tool. `input` (optional) validates and shapes the
- *  args; `handler` DESCRIBES the work against the live surface `client` and the
- *  adapter runs it at the one request edge; `description` is the tool's
- *  `tools/list` blurb and `title` its display name — MCP's two distinct
- *  metadata fields, the first written for the model choosing the tool, the
- *  second for the human reading a host's tool list. `title` is optional and a
- *  host that has none falls back to `name`, which is the machine spelling
- *  (`lifecycle_sendInput`) rather than a phrase.
- *
- *  `mutates` flags the tool for host authz (`readOnlyHint`/`destructiveHint`).
- *  It is OPTIONAL but defaults CONSERVATIVELY: an absent `mutates` is treated as
- *  MUTATING (`destructiveHint: true`), because `readOnlyHint: true` can let an MCP
- *  host auto-execute a tool unconfirmed — so an unannotated tool must fail SAFE
- *  (assume it writes), never silently advertise as a harmless read. Declare
- *  `mutates: false` ONLY for a genuinely read-only tool (a conscious, reviewable
- *  opt-in into the auto-approvable hint). */
-export interface BespokeTool<I = unknown, O = unknown> {
-  input?: ToolInputSchema<I>;
-  mutates?: boolean;
-  description?: string;
-  title?: string;
-  handler: (
-    args: I,
-    // The surface client is consumer-typed; the adapter holds it opaquely.
-    // biome-ignore lint/suspicious/noExplicitAny: client shape is the consumer's, opaque here.
-    client: any,
-    /** The MCP request's own signal, for a handler that must hand one to a
-     *  scaffold speaking `AbortSignal`. Ignore it otherwise — see the module
-     *  doc: the handler's effect is already run under this signal. */
-    signal: AbortSignal | undefined,
-  ) => Effect.Effect<O, unknown>;
-}
+import { wrapValue } from "@kolu/surface/verbs";
 
 /** The MCP `CallTool` result shape we emit. `content` is prose for the model to
  *  read; `structuredContent` is data for the caller to act on, so neither side
